src/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "dinder's backend (FastAPI and Pydantic)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
pydantic = ">=2.10.5,<3.0.0"
fastapi = ">=0.115.7,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
asyncpg = ">=0.30.0,<0.31.0"
python-dotenv = ">=1.0.1,<2.0.0"
supabase = ">=2.11.0,<3.0.0"
fastapi-cors = "^0.0.6"
fastembed = "^0.6.0"
pandas = "^2.2.3"
pytest-asyncio = "^0.25.3"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.4,<9.0.0"
isort = "^5.13.2"
black = "^24.8.0"
mypy = "^1.11.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
